+++ Movies-3rd-Challenge/NetworkService/NetworkService.swift
@@ -5,7 +5,8 @@ class NetworkService {
     private let searchMoviesURLString = "https://api.kinopoisk.dev/v1.4/movie/search"
     private let universalSearchMoviesURLString = "https://api.kinopoisk.dev/v1.4/movie"
     private let movieDetailURLString = "https://api.kinopoisk.dev/v1.4/movie/"
-    private let apiKey = Secrets.apiKey
+    private let apiKey = "83FBG6T-WSTMKQY-K4H377M-DWEBZG6"
+
     
     static let shared = NetworkService()
